requirements.txt
Flask==2.0.1
requests==2.28.1
openmeteo-requests==1.1.0
requests-cache==1.1.0
retry-requests==2.0.0
pandas==1.5.3
pillow==10.4.0 